Abstract eth wrapper behavior in integration tests
